requirements.txt
# Core dependencies
pydantic>=2.0
python-dateutil>=2.8.2

# LLM and Agent Framework
langgraph>=0.0.15
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-google-genai>=1.0.0

# Document Processing
pdfplumber>=0.9.0
pillow>=9.0.0
pytesseract>=0.3.10
python-docx>=0.8.11

# Fuzzy Matching
rapidfuzz>=3.0.0

# Image Processing
opencv-python>=4.8.0
numpy>=1.24.0

# API Framework (optional)
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# UI Framework (optional)
streamlit>=1.28.0
pandas>=2.0.0

# Utilities
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
